requirements.txt
fastapi
uvicorn
google-generativeai
openai
anthropic
yt-dlp
chromadb
sentence-transformers
python-dotenv
